+++ docs/source/conf.py
@@ -5,9 +5,9 @@ import sys
 sys.path.insert(0, os.path.abspath("../../src"))
 
 project = "openai-structured"
-copyright = "2024, Yaniv Golan"
+copyright = "2025, Yaniv Golan"
 author = "Yaniv Golan"
-version = "0.1.0"
+version = "0.3.0"
 
 extensions = [
     "sphinx.ext.autodoc",